Changed that all components that depend on each other are in one widget
